Modernize the C style of those autoconf tests.
The intent is to produce fewer warnings when configuring with -verbose.
Note that the warning on "implicit declaration of function" remains,
for relatively good reasons.


git-svn-id: http://caml.inria.fr/svn/ocaml/branches/cc-optim@16330 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ config/auto-aux/sizes.c
@@ -16,7 +16,10 @@
 int main(int argc, char **argv)
 {
   printf("%d %d %d %d %d\n",
-         sizeof(int), sizeof(long), sizeof(long *), sizeof(short),
-         sizeof(long long));
+         (int) sizeof(int),
+         (int) sizeof(long),
+         (int) sizeof(long *),
+         (int) sizeof(short),
+         (int) sizeof(long long));
   return 0;
 }
